Textbox kontrolleri eklendi.

+++ FrmBolumler.cs
@@ -52,52 +52,80 @@ namespace YurtKayitSistemi
 
         private void FrmBolumler_Load(object sender, EventArgs e)
         {
-           //  TODO: This line of code loads data into the 'yurtOtomasyonuDataSet1.Bolumler' table. You can move, or remove it, as needed.
+            //  TODO: This line of code loads data into the 'yurtOtomasyonuDataSet1.Bolumler' table. You can move, or remove it, as needed.
             this.bolumlerTableAdapter.Fill(this.yurtOtomasyonuDataSet1.Bolumler);
 
         }
 
         private void imgAdd_Click(object sender, EventArgs e)
         {
-            try
-            {
-                
-                SqlCommand komut1 = new SqlCommand("insert into Bolumler (BolumAd) values (@p1)", bgl.baglanti());
-                komut1.Parameters.AddWithValue("@p1", txtBolumAd.Text);
-                komut1.ExecuteNonQuery();
-                bgl.baglanti().Close();
-                MessageBox.Show("Bölüm başarıyla eklendi.","Bilgi",MessageBoxButtons.OK,MessageBoxIcon.Information);
-                this.bolumlerTableAdapter.Fill(this.yurtOtomasyonuDataSet1.Bolumler);
-            }
-            catch (Exception ex)
-            {
-                System.Diagnostics.Debug.WriteLine(ex.Message);
-                MessageBox.Show("Bölüm eklenirken bir hata oluştu!", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Error);
-            }
-            
+
+            //SELECT VE INSERT SORUSU !!!
+
+            //SqlCommand cmd = new SqlCommand("select BolumAd from Bolumler where BolumAd=@k1",bgl.baglanti());
+            //cmd.Parameters.AddWithValue("@k1", txtBolumAd.Text);
+            //SqlDataReader dr = cmd.ExecuteReader();
+            //while (dr.Read())
+            //{
+            //    if (dr[0].ToString() == txtBolumAd.Text)
+            //    {
+            //        MessageBox.Show("Test");
+            //    }
+            //    else
+            //    {
+            //        SqlCommand komut1 = new SqlCommand("insert into Bolumler (BolumAd) values (@k1)", bgl.baglanti());
+            //        komut1.Parameters.AddWithValue("@k1", txtBolumAd.Text);
+            //        komut1.ExecuteNonQuery();
+            //        bgl.baglanti().Close();
+            //        MessageBox.Show("Bölüm başarıyla eklendi.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            //        this.bolumlerTableAdapter.Fill(this.yurtOtomasyonuDataSet1.Bolumler);
+            //    }
+            //}
+            //bgl.baglanti().Close();
+            //dr.Close();
+            SqlCommand komut1 = new SqlCommand("insert into Bolumler (BolumAd) values (@k1)", bgl.baglanti());
+            komut1.Parameters.AddWithValue("@k1", txtBolumAd.Text);
+            komut1.ExecuteNonQuery();
+            bgl.baglanti().Close();
+            MessageBox.Show("Bölüm başarıyla eklendi.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            this.bolumlerTableAdapter.Fill(this.yurtOtomasyonuDataSet1.Bolumler);
+
+
+
+
+
         }
 
         private void imgDelete_Click(object sender, EventArgs e)
         {
-            try
+            if (txtBolumAd.Text == "")
             {
-               
-                SqlCommand komut2 = new SqlCommand("delete from Bolumler where Bolumid=@p1", bgl.baglanti());
-                komut2.Parameters.AddWithValue("@p1", txtBolumId.Text);
-                komut2.ExecuteNonQuery();
-                bgl.baglanti().Close();
-                MessageBox.Show("Bölüm başarıyla silindi!", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                this.bolumlerTableAdapter.Fill(this.yurtOtomasyonuDataSet1.Bolumler);
-                txtBolumAd.Text = "";
-                txtBolumId.Text = "";
+                MessageBox.Show("Bölüm adı boş geçilemez.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Warning);
             }
-            catch (Exception ex)
+            else if (txtBolumId.Text == "")
             {
-                System.Diagnostics.Debug.WriteLine(ex.Message);
-                MessageBox.Show("Bölüm silinirken bir hata oluştu!", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                MessageBox.Show("Bölüm ID boş geçilemez.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Warning);
             }
-           
+            else
+            {
+                try
+                {
 
+                    SqlCommand komut2 = new SqlCommand("delete from Bolumler where Bolumid=@p1", bgl.baglanti());
+                    komut2.Parameters.AddWithValue("@p1", txtBolumId.Text);
+                    komut2.ExecuteNonQuery();
+                    bgl.baglanti().Close();
+                    MessageBox.Show("Bölüm başarıyla silindi!", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    this.bolumlerTableAdapter.Fill(this.yurtOtomasyonuDataSet1.Bolumler);
+                    txtBolumAd.Text = "";
+                    txtBolumId.Text = "";
+                }
+                catch (Exception ex)
+                {
+                    System.Diagnostics.Debug.WriteLine(ex.Message);
+                    MessageBox.Show("Bölüm silinirken bir hata oluştu!", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
+            }
         }
 
        
@@ -117,28 +145,37 @@ namespace YurtKayitSistemi
 
         private void imgEdit_Click(object sender, EventArgs e)
         {
-            try
+            if (txtBolumAd.Text=="")
             {
-                
-                SqlCommand komut2 = new SqlCommand("update Bolumler SET BolumAd = @p1 where Bolumid = @p2", bgl.baglanti());
-                komut2.Parameters.AddWithValue("p2", txtBolumId.Text);
-                komut2.Parameters.AddWithValue("p1", txtBolumAd.Text);
-                komut2.ExecuteNonQuery();
-                bgl.baglanti().Close();
-                MessageBox.Show("Bölüm adı başarıyla güncellendi.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                this.bolumlerTableAdapter.Fill(this.yurtOtomasyonuDataSet1.Bolumler);
-                txtBolumAd.Text = "";
-                txtBolumId.Text = "";
+                MessageBox.Show("Bölüm adı boş geçilemez.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Warning);
             }
-            catch (Exception ex)
+            else if(txtBolumId.Text=="")
             {
-
-                System.Diagnostics.Debug.WriteLine(ex.Message);
-                MessageBox.Show("Bölüm adı güncellenirken bir hata oluştu!", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Error);
-
+                MessageBox.Show("Bölüm ID boş geçilemez.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Warning);
             }
-            
+            else
+            {
+                try
+                {
 
+                    SqlCommand komut2 = new SqlCommand("update Bolumler SET BolumAd = @p1 where Bolumid = @p2", bgl.baglanti());
+                    komut2.Parameters.AddWithValue("p2", txtBolumId.Text);
+                    komut2.Parameters.AddWithValue("p1", txtBolumAd.Text);
+                    komut2.ExecuteNonQuery();
+                    bgl.baglanti().Close();
+                    MessageBox.Show("Bölüm adı başarıyla güncellendi.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    this.bolumlerTableAdapter.Fill(this.yurtOtomasyonuDataSet1.Bolumler);
+                    txtBolumAd.Text = "";
+                    txtBolumId.Text = "";
+                }
+                catch (Exception ex)
+                {
+
+                    System.Diagnostics.Debug.WriteLine(ex.Message);
+                    MessageBox.Show("Bölüm adı güncellenirken bir hata oluştu!", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Error);
+
+                }
+            }
 
         }
     }
